stop using noStoreFetch
API is now correctly set up with `must-revalidate`, which tells the browser to ignore the disk cache and fetch fresh data after the max-age specified on the route.

The api subdomain now also correctly sets the Browser TTL setting to "Respect origin TTL" to remove that horrible 4 hour cache.

Same changes as done to bfstats-api.